This commit contains several important changes:
1) Corrections to device tree (in previous untagged commit)
2) Corrections to the totd_40mhz and mops_40mhz triggers.  Also make corresponding updates to the
integral_40mhz and deconvolve_40mhz modules.  Buffering was added to enable_40 signal.  Pipelining was
improved in integral_40mhz and deconvolv_40mhz.   A delay was added to properly synchronize the integral
and occupancy values in totd_40mhz.  The integration algorithm was modified to remove the 5us look-behind and
replace it with a running baseline calculation.  This now allows direct validation of the triggers in almost
all cases. (The only known exception is when signal just before the beginning of the trace has "pumped"
up the integral value.)  Integral is now capped at 2x the specified threshold in ToTD and MoPS, and resets
3us after passing the specified threshold.  Also the integral is now decayed with a linear decay of 1/2
count per 25ns time bin rather than an exponential decay to return to baseline faster.